✨ RecDotGov Tours + Tickets + Timed Entry (#189)
* ✨ RecDotGov Tours + Timed Entry (#151)

* Add support for tours (ticket and timed-entry) as new providers

A tour facility is currently modeled as a campground and each individual
tour is a campsite. User interface was not updated yet but the logic
does the work as expected.

* Add RecreationDotGovDailyTicket and RecreationDotGovDailyTimedEntry

With this endpoint, available time slots can be extracted. They're
attached as equipments on each tour (as campsite).

* RecreationDotGov Tours (#1)

Flake8 + ABC

* Fix Flake8 lint errors

Co-authored-by: Justin Flannery <juftin@juftin.com>

* ✨ RecDotGov Tours + Timed Entry

* 📝 New Provider Docs (#190)

* 📝 New Provider Docs

* 📝 Provider Docs

Co-authored-by: jingyuanliang <jingyuanliang@google.com>